rename `standard` license detector to `local`
