+++ internal/cli/cmd/app/app.go
@@ -28,7 +28,7 @@ import (
 	"github.com/CycloneDX/cyclonedx-gomod/internal/sbom"
 	"github.com/CycloneDX/cyclonedx-gomod/pkg/generate/app"
 	"github.com/CycloneDX/cyclonedx-gomod/pkg/licensedetect"
-	"github.com/CycloneDX/cyclonedx-gomod/pkg/licensedetect/standard"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/licensedetect/local"
 )
 
 func New() *ffcli.Command {
@@ -101,7 +101,7 @@ func Exec(options Options) error {
 
 	var licenseDetector licensedetect.Detector
 	if options.ResolveLicenses {
-		licenseDetector = standard.NewDetector(logger)
+		licenseDetector = local.NewDetector(logger)
 	}
 
 	generator, err := app.NewGenerator(options.ModuleDir,
