+++ pkg/generate/app/generator_test.go
@@ -31,7 +31,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/CycloneDX/cyclonedx-gomod/internal/testutil"
-	"github.com/CycloneDX/cyclonedx-gomod/pkg/licensedetect/standard"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/licensedetect/local"
 )
 
 func TestNewGenerator(t *testing.T) {
@@ -63,7 +63,7 @@ func TestGenerator_Generate(t *testing.T) {
 		fixturePath := testutil.ExtractFixtureArchive(t, "../testdata/simple.tar.gz")
 
 		g, err := NewGenerator(fixturePath,
-			WithLicenseDetector(standard.NewDetector(zerolog.Nop())),
+			WithLicenseDetector(local.NewDetector(zerolog.Nop())),
 			WithLogger(testutil.SilentLogger),
 			WithIncludeStdlib(true))
 		require.NoError(t, err)
@@ -88,7 +88,7 @@ func TestGenerator_Generate(t *testing.T) {
 			WithIncludeFiles(true),
 			WithIncludePackages(true),
 			WithIncludeStdlib(true),
-			WithLicenseDetector(standard.NewDetector(zerolog.Nop())),
+			WithLicenseDetector(local.NewDetector(zerolog.Nop())),
 			WithLogger(testutil.SilentLogger))
 		require.NoError(t, err)
 
@@ -111,7 +111,7 @@ func TestGenerator_Generate(t *testing.T) {
 		g, err := NewGenerator(fixturePath,
 			WithIncludePackages(true),
 			WithIncludeStdlib(true),
-			WithLicenseDetector(standard.NewDetector(zerolog.Nop())),
+			WithLicenseDetector(local.NewDetector(zerolog.Nop())),
 			WithLogger(testutil.SilentLogger))
 		require.NoError(t, err)
 
@@ -133,7 +133,7 @@ func TestGenerator_Generate(t *testing.T) {
 
 		g, err := NewGenerator(fixturePath,
 			WithIncludeStdlib(true),
-			WithLicenseDetector(standard.NewDetector(zerolog.Nop())),
+			WithLicenseDetector(local.NewDetector(zerolog.Nop())),
 			WithLogger(testutil.SilentLogger),
 			WithMainDir("cmd/purl"))
 		require.NoError(t, err)
@@ -156,7 +156,7 @@ func TestGenerator_Generate(t *testing.T) {
 
 		g, err := NewGenerator(fixturePath,
 			WithIncludeStdlib(true),
-			WithLicenseDetector(standard.NewDetector(zerolog.Nop())),
+			WithLicenseDetector(local.NewDetector(zerolog.Nop())),
 			WithLogger(testutil.SilentLogger),
 			WithMainDir("cmd/uuid"))
 		require.NoError(t, err)
@@ -179,7 +179,7 @@ func TestGenerator_Generate(t *testing.T) {
 
 		g, err := NewGenerator(fixturePath,
 			WithIncludeStdlib(true),
-			WithLicenseDetector(standard.NewDetector(zerolog.Nop())),
+			WithLicenseDetector(local.NewDetector(zerolog.Nop())),
 			WithLogger(testutil.SilentLogger))
 		require.NoError(t, err)
 
@@ -203,7 +203,7 @@ func TestGenerator_Generate(t *testing.T) {
 			WithIncludeFiles(true),
 			WithIncludePackages(true),
 			WithIncludeStdlib(true),
-			WithLicenseDetector(standard.NewDetector(zerolog.Nop())),
+			WithLicenseDetector(local.NewDetector(zerolog.Nop())),
 			WithLogger(testutil.SilentLogger))
 		require.NoError(t, err)
 
@@ -226,7 +226,7 @@ func TestGenerator_Generate(t *testing.T) {
 		g, err := NewGenerator(fixturePath,
 			WithIncludePackages(true),
 			WithIncludeStdlib(true),
-			WithLicenseDetector(standard.NewDetector(zerolog.Nop())),
+			WithLicenseDetector(local.NewDetector(zerolog.Nop())),
 			WithLogger(testutil.SilentLogger))
 		require.NoError(t, err)
