+++ pkg/generate/app/options_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/require"
 
-	"github.com/CycloneDX/cyclonedx-gomod/pkg/licensedetect/standard"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/licensedetect/local"
 )
 
 func TestWithIncludeFiles(t *testing.T) {
@@ -49,7 +49,7 @@ func TestWithIncludeStdlib(t *testing.T) {
 }
 
 func TestWithLicenseDetector(t *testing.T) {
-	detector := standard.NewDetector(zerolog.Nop())
+	detector := local.NewDetector(zerolog.Nop())
 
 	g := &generator{licenseDetector: detector}
 	err := WithLicenseDetector(detector)(g)
