+++ pkg/licensedetect/local/detector_test.go
@@ -15,7 +15,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) OWASP Foundation. All Rights Reserved.
 
-package standard
+package local
 
 import (
 	"testing"
